Change Vector's binary search return type and change size and capacity type to uint64_t; add push_sorted method to Vector; fix sort::is_sorted accepting uint16_t instead of uint64_t as a parameter type

+++ SNWK_Core/include/Utils/Sort.h
@@ -117,7 +117,7 @@ void quicksort (T* arr, uint64_t size)
 }
 
 template <typename T>
-bool is_sorted (const T* arr, uint16_t size)
+bool is_sorted (const T* arr, uint64_t size)
 {
     for (uint64_t i = 1; i < size; i++)
     {
